requirements.txt
Django==2.0.7
django-filter==1.1.0
django-tables2==1.21.2
django-widget-tweaks==1.4.2
psycopg2-binary==2.7.5
pytz==2018.5
